refactor: draw the network component with *iced*

+++ src/config.rs
@@ -94,8 +94,8 @@ pub struct NetworkGraphContainerProps {
 
 #[derive(Deserialize)]
 pub struct NetworkProps {
-  pub update_interval: u32,
-  pub public_ip_update_interval: Option<u32>,
+  pub update_interval: u64,
+  pub public_ip_update_interval: Option<u64>,
   pub interface_regex: SerializableRegex,
   pub graphs: NetworkGraphContainerProps,
 }
